CacheConnection R6 Class

An R6 class for handling RDS caching for data analysis and report generation.

Methods

Public methods

CacheConnection$new()

CacheConnection$load_from_disk()

CacheConnection$save_to_disk()

CacheConnection$append_page_note()

CacheConnection$get_notes()

CacheConnection$reactive()

CacheConnection$get_cache_path()

CacheConnection$set_cache_path()

CacheConnection$get_data()

CacheConnection$get_country()

CacheConnection$get_country_iso()

CacheConnection$get_indicator_groups()

CacheConnection$get_vaccine_indicators()

CacheConnection$get_data_with_excluded_years()

CacheConnection$get_adjusted_data()

CacheConnection$get_threshold()

CacheConnection$get_excluded_years()

CacheConnection$get_k_factors()

CacheConnection$get_adjusted_flag()

CacheConnection$get_start_survey_year()

CacheConnection$get_survey_estimates()

CacheConnection$get_national_estimates()

CacheConnection$get_un_estimates()

CacheConnection$get_wuenic_estimates()

CacheConnection$get_national_survey()

CacheConnection$get_regional_survey()

CacheConnection$get_wiq_survey()

CacheConnection$get_area_survey()

CacheConnection$get_education_survey()

CacheConnection$get_map_mapping()

CacheConnection$get_survey_mapping()

CacheConnection$get_page_notes()

CacheConnection$set_data()

CacheConnection$set_adjusted_data()

CacheConnection$set_threshold()

CacheConnection$set_excluded_years()

CacheConnection$set_k_factors()

CacheConnection$toggle_adjusted_flag()

CacheConnection$set_start_survey_year()

CacheConnection$set_survey_estimates()

CacheConnection$set_national_estimates()

CacheConnection$set_un_estimates()

CacheConnection$set_wuenic_estimates()

CacheConnection$set_national_survey()

CacheConnection$set_regional_survey()

CacheConnection$set_wiq_survey()

CacheConnection$set_area_survey()

CacheConnection$set_education_survey()

CacheConnection$set_survey_mapping()

CacheConnection$set_map_mapping()

CacheConnection$clone()

Method new()

Usage

CacheConnection$new(rds_path = NULL, countdown_data = NULL)

Arguments

rds_path

Path to the RDS file (NULL for in-memory only).

countdown_data

Countdown data (class cd_data). Load data from RDS file

Method load_from_disk()

Usage

CacheConnection$load_from_disk()

Returns

None Save data to disk (only if changed and RDS path is not NULL)

Method save_to_disk()

Usage

CacheConnection$save_to_disk()

Returns

None Reactive method for the CacheConnection object

Method append_page_note()

Usage

CacheConnection$append_page_note(
  page_id,
  object_id,
  note,
  title,
  parameters = list(),
  include_in_report = FALSE,
  include_plot_table = FALSE,
  single_entry = FALSE
)

Method get_notes()

Usage

CacheConnection$get_notes(page_id, object_id, parameters = NULL)

Method reactive()

Usage

CacheConnection$reactive()

Returns

A Shiny reactive expression tracking changes Gets for count down data

Method get_cache_path()

Usage

CacheConnection$get_cache_path()

Method set_cache_path()

Usage

CacheConnection$set_cache_path(value)

Method get_data()

Usage

CacheConnection$get_data()

Returns

a tibble of class cd_data. Gets for count down data with excluded years

Method get_country()

Usage

CacheConnection$get_country()

Method get_country_iso()

Usage

CacheConnection$get_country_iso()

Method get_indicator_groups()

Usage

CacheConnection$get_indicator_groups()

Method get_vaccine_indicators()

Usage

CacheConnection$get_vaccine_indicators()

Method get_data_with_excluded_years()

Usage

CacheConnection$get_data_with_excluded_years()

Returns

a tibble of class cd_data. Gets for count down data

Method get_adjusted_data()

Usage

CacheConnection$get_adjusted_data()

Returns

a tibble of class cd_data. Get performance threshold

Method get_threshold()

Usage

CacheConnection$get_threshold()

Returns

Numeric threshold Get excluded years

Method get_excluded_years()

Usage

CacheConnection$get_excluded_years()

Returns

Numeric vector of excluded years Get k-factors

Method get_k_factors()

Usage

CacheConnection$get_k_factors()

Returns

List of k-factors Get adjusted flag

Method get_adjusted_flag()

Usage

CacheConnection$get_adjusted_flag()

Returns

Factor TRUE or FALSE Get excluded years

Method get_start_survey_year()

Usage

CacheConnection$get_start_survey_year()

Returns

Numeric vector of excluded years Get survey estimates

Method get_survey_estimates()

Usage

CacheConnection$get_survey_estimates()

Returns

List of survey estimates Get national estimates

Method get_national_estimates()

Usage

CacheConnection$get_national_estimates()

Returns

List of national estimates Get mapping information

Method get_un_estimates()

Usage

CacheConnection$get_un_estimates()

Returns

List of mapping information Get mapping information

Method get_wuenic_estimates()

Usage

CacheConnection$get_wuenic_estimates()

Returns

List of mapping information Get mapping information

Method get_national_survey()

Usage

CacheConnection$get_national_survey()

Returns

List of mapping information Get mapping information

Method get_regional_survey()

Usage

CacheConnection$get_regional_survey()

Returns

List of mapping information Get mapping information

Method get_wiq_survey()

Usage

CacheConnection$get_wiq_survey()

Returns

List of mapping information Get mapping information

Method get_area_survey()

Usage

CacheConnection$get_area_survey()

Returns

List of mapping information Get mapping information

Method get_education_survey()

Usage

CacheConnection$get_education_survey()

Returns

List of mapping information Get mapping information

Method get_map_mapping()

Usage

CacheConnection$get_map_mapping()

Returns

List of mapping information Get mapping information

Method get_survey_mapping()

Usage

CacheConnection$get_survey_mapping()

Returns

List of mapping information Get page notes

Method get_page_notes()

Usage

CacheConnection$get_page_notes()

Returns

Tibble of page notes Set countdown data

Method set_data()

Usage

CacheConnection$set_data(value)

Arguments

value

Countdown data to set

Returns

None Set adjusted countdown data

Method set_adjusted_data()

Usage

CacheConnection$set_adjusted_data(value)

Arguments

value

Countdown data to set

Returns

None Set performance threshold

Method set_threshold()

Usage

CacheConnection$set_threshold(value)

Arguments

value

Numeric threshold to set

Returns

None Set excluded years

Method set_excluded_years()

Usage

CacheConnection$set_excluded_years(value)

Arguments

value

Numeric vector of years to exclude

Returns

None Set k-factors

Method set_k_factors()

Usage

CacheConnection$set_k_factors(value)

Arguments

value

List of k-factors to set

Returns

None Toggle adjusted flag

Method toggle_adjusted_flag()

Usage

CacheConnection$toggle_adjusted_flag(adjusted)

Arguments

adjusted

Logical value for adjusted flag

Returns

None Set excluded years

Method set_start_survey_year()

Usage

CacheConnection$set_start_survey_year(value)

Arguments

value

Numeric vector of years to exclude

Returns

None Set survey estimates

Method set_survey_estimates()

Usage

CacheConnection$set_survey_estimates(value)

Arguments

value

List of survey estimates

Returns

None Set national estimates

Method set_national_estimates()

Usage

CacheConnection$set_national_estimates(value)

Arguments

value

List of national estimates

Returns

None Set UN estimates information

Method set_un_estimates()

Usage

CacheConnection$set_un_estimates(value)

Arguments

value

A data frame of mapping information

Returns

None Set WUENIC information

Method set_wuenic_estimates()

Usage

CacheConnection$set_wuenic_estimates(value)

Arguments

value

A data frame of mapping information

Returns

None Set national survey information

Method set_national_survey()

Usage

CacheConnection$set_national_survey(value)

Arguments

value

A data frame of mapping information

Returns

None Set subnational survey mapping information

Method set_regional_survey()

Usage

CacheConnection$set_regional_survey(value)

Arguments

value

A data frame of mapping information

Returns

None Set Wealth quintile information

Method set_wiq_survey()

Usage

CacheConnection$set_wiq_survey(value)

Arguments

value

A data frame of mapping information

Returns

None Set survey mapping information

Method set_area_survey()

Usage

CacheConnection$set_area_survey(value)

Arguments

value

A data frame of mapping information

Returns

None Set survey mapping information

Method set_education_survey()

Usage

CacheConnection$set_education_survey(value)

Arguments

value

A data frame of mapping information

Returns

None Set survey mapping information

Method set_survey_mapping()

Usage

CacheConnection$set_survey_mapping(value)

Arguments

value

A data frame of mapping information

Returns

None Set map mapping information

Method set_map_mapping()

Usage

CacheConnection$set_map_mapping(value)

Arguments

value

List of mapping information

Returns

None Update a field (with change tracking)

Method clone()

The objects of this class are cloneable with this method.

Usage

CacheConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.