Analyze National Immunization Coverage Data

analyze_coverage calculates and combines national immunization coverage data for a specified country, indicator, and denominator. It integrates local survey data and WUENIC (WHO-UNICEF estimates) and processes the provided data source, which is precomputed as a cd_indicator_coverage data frame.

Usage

analyze_coverage(
  .data,
  admin_level = c("national", "adminlevel_1", "district"),
  indicator = c("bcg", "anc1", "pcv3", "opv1", "opv2", "opv3", "penta2", "pcv1", "pcv2",
    "penta1", "penta3", "measles1", "rota1", "rota2", "instdeliveries", "measles2",
    "ipv1", "ipv2", "undervax", "dropout_penta13", "zerodose", "dropout_measles12",
    "dropout_penta3mcv1"),
  denominator = c("dhis2", "anc1", "penta1"),
  survey_data,
  wuenic_data,
  subnational_map = NULL,
  region = NULL
)

Arguments

.data

A precomputed data frame of type cd_indicator_coverage containing immunization coverage data. This dataset should include calculated metrics such as coverage indicators.

admin_level

Character. Specifies the administrative level for analysis. Options include: national, adminlevel_1 and district. Default is 'national'.

indicator

Character. The immunization indicator to analyze. Must be one of the supported values: bcg, anc1, pcv3, opv1, opv2, opv3, penta2, penta1, penta3, measles1, measles2, rota1, rota2, ipv1, ipv2, undervax, dropout_penta13, zerodose, anddropout_measles12, dropout_penta3mcv1

denominator

Character. The denominator used for calculating coverage estimates. Must be one of the following: dhis2, anc1, andpenta1.

survey_data

A data frame containing survey estimates for immunization coverage.

wuenic_data

A data frame containing WUENIC (WHO-UNICEF estimates) data for immunization coverage. This should include columns specific to WUENIC-based coverage estimates.

subnational_map

A data frame used to link subnational regions (e.g., districts or administrative level 1 areas) to their respective parent regions. This is required for subnational analyses. Default is NULL.

region

Character. Specifies the region of interest for subnational analyses. Required if admin_level is not 'national'. Default is NULL.

Value

A data frame of class cd_coverage, containing year-wise immunization coverage estimates from DHIS2, WUENIC, and survey data, along with confidence intervals where available. The resulting table is formatted for further analysis or visualization, with metadata describing the chosen administrative level, region, and denominator.

Details

This function provides a flexible way to analyze immunization coverage by combining data from multiple sources (survey data, WUENIC, and DHIS2). It supports analyses at different administrative levels and is designed for integration into dashboards, reports, or further data exploration workflows.

Examples

if (FALSE) { # \dontrun{
analyze_coverage(
  .data = precomputed_data,
  admin_level = 'national',
  indicator = 'penta3',
  denominator = 'penta1',
  survey_data = survey_df,
  wuenic_data = wuenic_df
)
} # }