Check for Missing Values by District and Year

calculate_district_completeness_summary calculates the percentage of districts with complete data (no missing values) for each year. By assessing missingness across all districts for each indicator, it provides a summary of data completeness on a per-district basis.

Usage

calculate_district_completeness_summary(.data)

Arguments

.data

A cd_data object containing indicator data with columns prefixed by mis_, indicating missing values for each district and year.

Value

A cd_district_completeness_summary object containing a tibble. This tibble summarizes the percentage of districts with no missing values for each indicator per year, as well as an overall yearly summary across all indicators.

Details

For each year, the function calculates the percentage of districts without missing values per indicator.

The output includes specific summary statistics for key indicator groups, such as vaccination indicators and tracer indicators, providing insights into data completeness at a granular level.

All percentages are rounded to two decimal places.

Examples

if (FALSE) { # \dontrun{
  # Calculate the percentage of districts with complete data per year
  calculate_district_completeness_summary(data)
} # }