CacheConnection R6 Class

An R6 class for handling RDS caching for data analysis and report generation.

Methods

Public methods

CacheConnection$new()

CacheConnection$load_from_disk()

CacheConnection$save_to_disk()

CacheConnection$append_page_note()

CacheConnection$get_notes()

CacheConnection$reactive()

CacheConnection$set_cache_path()

CacheConnection$set_countdown_data()

CacheConnection$set_adjusted_data()

CacheConnection$set_performance_threshold()

CacheConnection$set_excluded_years()

CacheConnection$set_k_factors()

CacheConnection$set_adjusted_flag()

CacheConnection$set_survey_estimates()

CacheConnection$set_national_estimates()

CacheConnection$set_start_survey_year()

CacheConnection$set_denominator()

CacheConnection$set_selected_admin_level_1()

CacheConnection$set_selected_district()

CacheConnection$set_mapping_years()

CacheConnection$set_un_estimates()

CacheConnection$set_wuenic_estimates()

CacheConnection$set_national_survey()

CacheConnection$set_regional_survey()

CacheConnection$set_wiq_survey()

CacheConnection$set_area_survey()

CacheConnection$set_education_survey()

CacheConnection$set_survey_mapping()

CacheConnection$set_map_mapping()

CacheConnection$clone()

Method new()

Usage

CacheConnection$new(rds_path = NULL, countdown_data = NULL)

Arguments

rds_path

Path to the RDS file (NULL for in-memory only).

countdown_data

Countdown data (class cd_data). Load data from RDS file

Method load_from_disk()

Usage

CacheConnection$load_from_disk()

Returns

None Save data to disk (only if changed and RDS path is not NULL)

Method save_to_disk()

Usage

CacheConnection$save_to_disk()

Returns

None Reactive method for the CacheConnection object

Method append_page_note()

Usage

CacheConnection$append_page_note(
  page_id,
  object_id,
  note,
  parameters = list(),
  include_in_report = FALSE,
  include_plot_table = FALSE,
  single_entry = FALSE
)

Method get_notes()

Usage

CacheConnection$get_notes(page_id, object_id = NULL, parameters = NULL)

Method reactive()

Usage

CacheConnection$reactive()

Returns

A Shiny reactive expression tracking changes Update a field (with change tracking)

Method set_cache_path()

Usage

CacheConnection$set_cache_path(value)

Method set_countdown_data()

Usage

CacheConnection$set_countdown_data(value)

Method set_adjusted_data()

Usage

CacheConnection$set_adjusted_data(value)

Method set_performance_threshold()

Usage

CacheConnection$set_performance_threshold(value)

Method set_excluded_years()

Usage

CacheConnection$set_excluded_years(value)

Method set_k_factors()

Usage

CacheConnection$set_k_factors(value)

Method set_adjusted_flag()

Usage

CacheConnection$set_adjusted_flag(value)

Method set_survey_estimates()

Usage

CacheConnection$set_survey_estimates(value)

Method set_national_estimates()

Usage

CacheConnection$set_national_estimates(value)

Method set_start_survey_year()

Usage

CacheConnection$set_start_survey_year(value)

Method set_denominator()

Usage

CacheConnection$set_denominator(value)

Method set_selected_admin_level_1()

Usage

CacheConnection$set_selected_admin_level_1(value)

Method set_selected_district()

Usage

CacheConnection$set_selected_district(value)

Method set_mapping_years()

Usage

CacheConnection$set_mapping_years(value)

Method set_un_estimates()

Usage

CacheConnection$set_un_estimates(value)

Method set_wuenic_estimates()

Usage

CacheConnection$set_wuenic_estimates(value)

Method set_national_survey()

Usage

CacheConnection$set_national_survey(value)

Method set_regional_survey()

Usage

CacheConnection$set_regional_survey(value)

Method set_wiq_survey()

Usage

CacheConnection$set_wiq_survey(value)

Method set_area_survey()

Usage

CacheConnection$set_area_survey(value)

Method set_education_survey()

Usage

CacheConnection$set_education_survey(value)

Method set_survey_mapping()

Usage

CacheConnection$set_survey_mapping(value)

Method set_map_mapping()

Usage

CacheConnection$set_map_mapping(value)

Method clone()

The objects of this class are cloneable with this method.

Usage

CacheConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.